requirements.txt
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1699370045702/work
asgiref==3.7.2
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1703154663129/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1695989934239/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1696001737800/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1701731166956/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1702406360642/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1691048088238/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1691009212940/work
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1675116244235/work
Django==5.0.2
django-s3-sqlite==0.0.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1701026962277/work
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1695536281965/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1695397382064/work
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1702310516100/work/libmambapy
mamba @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1702310516100/work/mamba
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1702317138933/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1701708255999/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1693086607691/work
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1696355969704/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
PyJWT==2.8.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pytz==2024.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
rest-framework-simplejwt==0.0.2
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1699007390921/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1695996997220/work
sqlparse==0.4.4
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1691671248568/work
truststore @ file:///home/conda/feedstock_root/build_artifacts/truststore_1694154605758/work
typing_extensions==4.9.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1699933488691/work
zstandard==0.22.0
